backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-ai[groq]
python-dotenv
